+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added ErrorKind::Locking to indicate problems acquiring synchronisation locks.
 - Added the ability to define families of L-Systems that share common tokens.
+- Path is now a fuller collection, with indexing and iterator support
 
 ### Changed
 - README updates.